langgraph_project/requirements.txt
langchain==0.1.0
langgraph==0.0.10
chromadb==0.4.22
python-dotenv==1.0.0
openai==1.12.0
tiktoken==0.5.2
pydantic==2.5.2
fastapi==0.109.0
uvicorn==0.27.0
pypdf==5.4.0
langchain-openai==0.3.15
langchain-community==0.3.23
langchain-chroma==0.1.0
fpdf==1.7.2
langchain-ollama==0.1.1
scikit-learn==1.4.1.post1 